// apps/frontend/src/shared/types/branch/branch.type.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Branch = { id: string, name: string | null, chatId: string, parentId: string | null, modelId: string | null, updatedAt: string, };

export type UpdateBranch = { id: string, name: string | null, modelId: string | null, };