// apps/frontend/src/shared/types/chat/chat.type.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { WorkspaceItemContextInheritanceMode } from "../workspace/workspace-item-context-inheritance-mode.type";

export type Chat = { id: string, name: string, workspaceId: string, parentId: string | null, currentBranchId: string | null, context: string | null, contextInheritanceMode: WorkspaceItemContextInheritanceMode, isArchived: boolean, archivedAt: string, jsonMetadata: string | null, jsonExtensions: string | null, updatedAt: string, };